+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This file contains the list of changes made to pyjoulescope_ui.
 *   Fixed plugin window instances become invalid #74.
 *   Improved logging for multiprocessing.
 *   Fixed downsampled JLS files display dropped samples as 0 #75.
+*   Added preference to elevate Windows process priority, enabled by default.
 
 
 ## 0.8.11